+++ packages/annotation_indexer/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Add `ListViewPath` class and `paths` list.
+- Add `GoRouter` class and `goRouter` instance.
+- Index target limit `.list_view.dart`
+
 ## 0.0.1
 
 - Initial version.